updating the user_calculator.py program to support float addition

+++ Demo_folder/user_calculator.py
@@ -1,6 +1,6 @@
 try:
-    a=int(input("Enter number 1 : "))
-    b=int(input("Enter number 2 : "))
+    a=float(input("Enter number 1 : "))
+    b=float(input("Enter number 2 : "))
 except Exception:
     print("Please enter a valid input. Entered input is not an integer\n")
 else:
